Moved processing luminosity function to plotting
