+++ alcor/services/plots/luminosity_function.py
@@ -1,22 +1,35 @@
 from math import (log10,
                   sqrt)
-from typing import Tuple
+from typing import (Iterable,
+                    Iterator,
+                    Tuple,
+                    List)
 
 import matplotlib
 
 # More info at
 # http://matplotlib.org/faq/usage_faq.html#what-is-a-backend for details
 # TODO: use this: https://stackoverflow.com/a/37605654/7851470
+from alcor.types import StarsBinsType
 
 matplotlib.use('Agg')
 from matplotlib import pyplot as plt
 import numpy as np
-from sqlalchemy.orm.session import Session
 
+from alcor.models import Star
 from alcor.models.luminosity_function import Point
-from alcor.services.common import FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME
-from alcor.services.data_access import fetch_all
+from alcor.services.common import (FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME,
+                                   stars_packer,
+                                   star_bolometric_indexer,
+                                   bolometric_indexer)
 
+MIN_BOLOMETRIC_MAGNITUDE = 6.
+STARS_BIN_SIZE = 0.5
+MAX_BOLOMETRIC_MAGNITUDE = 21.
+# More info at (The white dwarf population within 40 pc of the Sun,
+# Torres et al., 2016)
+SMALLEST_ERROR_BARS_BINS_INDEXES = {15, 16, 17}
+OBSERVATIONAL_DATA_TRUSTED_BINS_OBJECT_COUNT = 220
 # Observational LF of 40pc sample from Althaus
 OBSERVATIONAL_AVG_BIN_MAGNITUDES = np.arange(7.75, 17.25, 0.5)
 OBSERVATIONAL_STARS_COUNTS = [3, 4, 5, 7, 12, 17, 17, 12, 20, 19, 37, 42, 52,
@@ -37,9 +50,15 @@ OBSERVATIONAL_LOWER_ERRORBARS = [
     for stars_count in OBSERVATIONAL_STARS_COUNTS]
 OBSERVATIONAL_ASYMMETRIC_ERRORBARS = [OBSERVATIONAL_LOWER_ERRORBARS,
                                       OBSERVATIONAL_UPPER_ERRORBARS]
+bolometric_index = bolometric_indexer(min_magnitude=MIN_BOLOMETRIC_MAGNITUDE,
+                                      stars_bin_size=STARS_BIN_SIZE)
+STARS_BINS_COUNT = bolometric_index(MAX_BOLOMETRIC_MAGNITUDE)
+star_bolometric_index = star_bolometric_indexer(bolometric_index)
+pack_stars = stars_packer(stars_bins_count=STARS_BINS_COUNT,
+                          star_bolometric_index=star_bolometric_index)
 
 
-def plot(session: Session,
+def plot(stars: List[Star],
          filename: str = 'luminosity_function.ps',
          figure_size: Tuple[float, float] = (7, 7),
          ratio: float = 10 / 13,
@@ -51,14 +70,16 @@ def plot(session: Session,
          marker: str = 's',
          capsize: float = 5,
          observational_line_color: str = 'r') -> None:
-    # TODO: Implement other fetching functions
-    graph_points = fetch_all(Point,
-                             session=session)
+    stars_bins = pack_stars(stars)
+    normalization_factor = stars_bins_normalization_factor(stars_bins)
+    stars_counts = map(len, stars_bins)
+    points = list(graph_points(stars_counts=stars_counts,
+                               normalization_factor=normalization_factor))
 
     avg_bin_magnitudes = [graph_point.avg_bin_magnitude
-                          for graph_point in graph_points]
+                          for graph_point in points]
     stars_count_logarithms = [graph_point.stars_count_logarithm
-                              for graph_point in graph_points]
+                              for graph_point in points]
 
     # As simulated number of white dwarfs (WDs) can be much greater than the
     # number of WDs for which observational luminosity function (LF) is
@@ -89,14 +110,6 @@ def plot(session: Session,
         except ValueError:
             # Some number so that errorbar would go below the plot
             lower_errorbars.append(5.0)
-
-    (avg_bin_magnitudes,
-     stars_count_logarithms,
-     upper_errorbars,
-     lower_errorbars) = zip(*sorted(zip(avg_bin_magnitudes,
-                                        stars_count_logarithms,
-                                        upper_errorbars,
-                                        lower_errorbars)))
 
     asymmetric_errorbars = [lower_errorbars,
                             upper_errorbars]
@@ -132,3 +145,62 @@ def plot(session: Session,
     subplot.set_aspect(ratio / subplot.get_data_ratio())
 
     plt.savefig(filename)
+
+
+def graph_points(*,
+                 stars_counts: Iterable[int],
+                 normalization_factor: float) -> Iterator[Point]:
+    for index, stars_count in enumerate(stars_counts):
+        if not stars_count:
+            continue
+
+        avg_bin_magnitude = (MIN_BOLOMETRIC_MAGNITUDE
+                             + STARS_BIN_SIZE * (index + 0.5))
+        logarithm = to_logarithm(stars_count=stars_count,
+                                 normalization_factor=normalization_factor)
+        upper_error_bar = to_upper_error_bar(stars_count)
+        lower_error_bar = to_lower_error_bar(stars_count)
+        yield Point(avg_bin_magnitude=avg_bin_magnitude,
+                    stars_count_logarithm=logarithm,
+                    upper_error_bar=upper_error_bar,
+                    lower_error_bar=lower_error_bar)
+
+
+def stars_bins_normalization_factor(
+        stars_bins: StarsBinsType,
+        stars_bins_indexes: Iterable[int] = SMALLEST_ERROR_BARS_BINS_INDEXES
+) -> float:
+    smallest_error_bars_bins_lengths_sum = (
+        sum(len(stars_bins[index])
+            for index in stars_bins_indexes))
+    return (FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME
+            * smallest_error_bars_bins_lengths_sum
+            / OBSERVATIONAL_DATA_TRUSTED_BINS_OBJECT_COUNT)
+
+
+def to_logarithm(stars_count: int,
+                 normalization_factor: float) -> float:
+    try:
+        return log10(stars_count / normalization_factor)
+    except ValueError:
+        return 0.
+
+
+def to_upper_error_bar(stars_count: int) -> float:
+    if stars_count == 0:
+        return 0.
+    else:
+        return (log10((stars_count + sqrt(stars_count))
+                      / FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME)
+                - log10(stars_count / FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME))
+
+
+def to_lower_error_bar(stars_count: int) -> float:
+    if stars_count == 0:
+        return 0.
+    elif stars_count == 1:
+        return 5.  # Some number so that error bar would go below the plot
+    else:
+        return (log10(stars_count / FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME)
+                - log10((stars_count - sqrt(stars_count))
+                        / FORTY_PARSEC_NORTHERN_HEMISPHERE_VOLUME))
