+++ alcor/services/processing/service.py
@@ -12,7 +12,6 @@ from alcor.services.data_access import (fetch_unprocessed_groups,
 def run(*,
         filtration_method: str,
         nullify_radial_velocity: bool,
-        w_luminosity_function: bool,
         w_velocities_clouds: bool,
         w_velocities_vs_magnitude: bool,
         w_lepine_criterion: bool,
@@ -33,7 +32,6 @@ def run(*,
                 group=group,
                 filtration_method=filtration_method,
                 nullify_radial_velocity=nullify_radial_velocity,
-                w_luminosity_function=w_luminosity_function,
                 w_velocities_clouds=w_velocities_clouds,
                 w_velocities_vs_magnitude=w_velocities_vs_magnitude,
                 w_lepine_criterion=w_lepine_criterion,
