+++ manage.py
@@ -108,9 +108,6 @@ def simulate(ctx: click.Context,
 @click.option('--nullify-radial-velocity', '-nrv',
               is_flag=True,
               help='Sets radial velocities to zero.')
-@click.option('--w-luminosity-function', '-lf',
-              is_flag=True,
-              help='Prepare data for plotting luminosity function.')
 @click.option('--w-velocities-clouds', '-uvw',
               is_flag=True,
               help='Prepare data for plotting velocity clouds.')
@@ -127,7 +124,6 @@ def process(ctx: click.Context,
             group_id: Optional[uuid.UUID],
             filtration_method: str,
             nullify_radial_velocity: bool,
-            w_luminosity_function: bool,
             w_velocities_clouds: bool,
             w_velocities_vs_magnitude: bool,
             w_lepine_criterion: bool,
@@ -148,7 +144,6 @@ def process(ctx: click.Context,
 
         processing.run(filtration_method=filtration_method,
                        nullify_radial_velocity=nullify_radial_velocity,
-                       w_luminosity_function=w_luminosity_function,
                        w_velocities_clouds=w_velocities_clouds,
                        w_velocities_vs_magnitude=w_velocities_vs_magnitude,
                        w_lepine_criterion=w_lepine_criterion,
